+++ tests/unittest/test_base_rickle.py
@@ -94,6 +94,7 @@ class TestBaseRickle(unittest.TestCase):
 
         with self.assertRaises(AttributeError):
             is_true = self.base_rickle.key_one == "value_one"
+
     def test_iterator(self):
         expected = {
             'first': {