+++ tests/unittest/test_rickle.py
@@ -94,8 +94,8 @@ class TestBaseRickle(unittest.TestCase):
         self.assertTrue(self.rickle.get("bowser").startswith("d061"))
 
 
-    def test_add_api_json(self):
-        self.rickle.add_api("api_result", "https://official-joke-api.appspot.com/random_joke")
+    def test_add_api(self):
+        self.rickle.add_api("api_result", "https://official-joke-api.appspot.com/random_joke", load_as_rick=True)
         keys = self.rickle.get("api_result").keys()
         self.assertTrue('type' in keys)
         self.assertTrue('setup' in keys)